+++ src/contexts/CountryContext/CountryContext.tsx
@@ -1,3 +1,4 @@
+/* eslint-disable @typescript-eslint/no-unused-vars */
 import React, { createContext, useMemo, useState, useEffect } from 'react';
 import AsyncStorage from '@react-native-async-storage/async-storage';
 import * as Localization from 'expo-localization';
@@ -6,11 +7,12 @@ import { countries, CountriesType } from '@styles';
 export const CountryContext = createContext<any>(null);
 
 const CountryProvider: React.FC = ({ children }) => {
-	const [country, setCountry] = useState<string>(countries[Localization.region! as CountriesType]);
+	const [country, setCountry] = useState<string>('');
+	// const [country, setCountry] = useState<string>(countries[Localization.region! as CountriesType]);
 
 	useEffect(() => {
 		const fetchCountry = async () => {
-			const storedLocation = await AsyncStorage.getItem('@country');
+			const storedLocation = await AsyncStorage.getItem('@77777');
 			if (storedLocation) setCountry(storedLocation);
 		};
 		fetchCountry();
@@ -18,7 +20,7 @@ const CountryProvider: React.FC = ({ children }) => {
 
 	useEffect(() => {
 		const storeCountry = async () => {
-			await AsyncStorage.setItem('@country', country!);
+			await AsyncStorage.setItem('@77777', country!);
 		};
 		storeCountry();
 	}, [country]);
